requirements.txt
lxml==4.3.4
numpy==1.16.4
opencv-python==4.1.0.25
Pillow==6.0.0
pytesseract==0.2.6
python-docx==0.8.10
